+++ packages/jest-core/src/getNoTestFound.ts
@@ -7,7 +7,7 @@
 
 import chalk = require('chalk');
 import type {Config} from '@jest/types';
-import pluralize from './pluralize';
+import {pluralize} from 'jest-util';
 import type {TestRunData} from './types';
 
 export default function getNoTestFound(