+++ packages/jest-runtime/src/index.ts
@@ -53,7 +53,12 @@ import type {MockMetadata, ModuleMocker} from 'jest-mock';
 import {escapePathForRegex} from 'jest-regex-util';
 import Resolver, {ResolveModuleConfig} from 'jest-resolve';
 import {EXTENSION as SnapshotExtension} from 'jest-snapshot';
-import {createDirectory, deepCyclicCopy} from 'jest-util';
+import {
+  createDirectory,
+  deepCyclicCopy,
+  invariant,
+  isNonNullable,
+} from 'jest-util';
 import {
   createOutsideJestVmPath,
   decodePossibleOutsideJestVmPath,
@@ -124,7 +129,7 @@ const NODE_MODULES = `${path.sep}node_modules${path.sep}`;
 const getModuleNameMapper = (config: Config.ProjectConfig) => {
   if (
     Array.isArray(config.moduleNameMapper) &&
-    config.moduleNameMapper.length
+    config.moduleNameMapper.length > 0
   ) {
     return config.moduleNameMapper.map(([regex, moduleName]) => ({
       moduleName,
@@ -208,6 +213,7 @@ export default class Runtime {
   private readonly esmConditions: Array<string>;
   private readonly cjsConditions: Array<string>;
   private isTornDown = false;
+  private isInsideTestCode: boolean | undefined;
 
   constructor(
     config: Config.ProjectConfig,
@@ -284,7 +290,7 @@ export default class Runtime {
     );
 
     if (config.automock) {
-      config.setupFiles.forEach(filePath => {
+      for (const filePath of config.setupFiles) {
         if (filePath.includes(NODE_MODULES)) {
           const moduleID = this._resolver.getModuleID(
             this._virtualMocks,
@@ -299,7 +305,7 @@ export default class Runtime {
           );
           this._transitiveShouldMock.set(moduleID, false);
         }
-      });
+      }
     }
 
     this.resetModules();
@@ -416,12 +422,15 @@ export default class Runtime {
     query = '',
   ): Promise<VMModule> {
     const cacheKey = modulePath + query;
+    const registry = this._isolatedModuleRegistry
+      ? this._isolatedModuleRegistry
+      : this._esmoduleRegistry;
 
     if (this._fileTransformsMutex.has(cacheKey)) {
       await this._fileTransformsMutex.get(cacheKey);
     }
 
-    if (!this._esmoduleRegistry.has(cacheKey)) {
+    if (!registry.has(cacheKey)) {
       invariant(
         typeof this._environment.getVmContext === 'function',
         'ES Modules are only supported if your test environment has the `getVmContext` function',
@@ -454,7 +463,7 @@ export default class Runtime {
           context,
         );
 
-        this._esmoduleRegistry.set(cacheKey, wasm);
+        registry.set(cacheKey, wasm);
 
         transformResolve();
         return wasm;
@@ -462,7 +471,7 @@ export default class Runtime {
 
       if (this._resolver.isCoreModule(modulePath)) {
         const core = this._importCoreModule(modulePath, context);
-        this._esmoduleRegistry.set(cacheKey, core);
+        registry.set(cacheKey, core);
 
         transformResolve();
 
@@ -526,11 +535,11 @@ export default class Runtime {
         }
 
         invariant(
-          !this._esmoduleRegistry.has(cacheKey),
+          !registry.has(cacheKey),
           `Module cache already has entry ${cacheKey}. This is a bug in Jest, please report it!`,
         );
 
-        this._esmoduleRegistry.set(cacheKey, module);
+        registry.set(cacheKey, module);
 
         transformResolve();
       } catch (error) {
@@ -539,7 +548,7 @@ export default class Runtime {
       }
     }
 
-    const module = this._esmoduleRegistry.get(cacheKey);
+    const module = registry.get(cacheKey);
 
     invariant(
       module,
@@ -562,15 +571,24 @@ export default class Runtime {
       // @ts-expect-error - exiting
       return;
     }
+    if (this.isInsideTestCode === false) {
+      throw new ReferenceError(
+        'You are trying to `import` a file outside of the scope of the test code.',
+      );
+    }
+
+    const registry = this._isolatedModuleRegistry
+      ? this._isolatedModuleRegistry
+      : this._esmoduleRegistry;
 
     if (specifier === '@jest/globals') {
-      const fromCache = this._esmoduleRegistry.get('@jest/globals');
+      const fromCache = registry.get('@jest/globals');
 
       if (fromCache) {
         return fromCache;
       }
       const globals = this.getGlobalsForEsm(referencingIdentifier, context);
-      this._esmoduleRegistry.set('@jest/globals', globals);
+      registry.set('@jest/globals', globals);
 
       return globals;
     }
@@ -586,7 +604,7 @@ export default class Runtime {
         return this.importMock(referencingIdentifier, specifier, context);
       }
 
-      const fromCache = this._esmoduleRegistry.get(specifier);
+      const fromCache = registry.get(specifier);
 
       if (fromCache) {
         return fromCache;
@@ -662,7 +680,7 @@ export default class Runtime {
         }
       }
 
-      this._esmoduleRegistry.set(specifier, module);
+      registry.set(specifier, module);
       return module;
     }
 
@@ -705,6 +723,11 @@ export default class Runtime {
       );
       process.exitCode = 1;
       return;
+    }
+    if (this.isInsideTestCode === false) {
+      throw new ReferenceError(
+        'You are trying to `import` a file outside of the scope of the test code.',
+      );
     }
 
     if (module.status === 'unlinked') {
@@ -766,10 +789,10 @@ export default class Runtime {
     const module = new SyntheticModule(
       [...cjsExports, 'default'],
       function () {
-        cjsExports.forEach(exportName => {
+        for (const exportName of cjsExports) {
           // @ts-expect-error: TS doesn't know what `this` is
           this.setExport(exportName, cjs[exportName]);
-        });
+        }
         // @ts-expect-error: TS doesn't know what `this` is
         this.setExport('default', cjs);
       },
@@ -804,10 +827,10 @@ export default class Runtime {
       const module = new SyntheticModule(
         Object.keys(invokedFactory),
         function () {
-          Object.entries(invokedFactory).forEach(([key, value]) => {
+          for (const [key, value] of Object.entries(invokedFactory)) {
             // @ts-expect-error: TS doesn't know what `this` is
             this.setExport(key, value);
-          });
+          }
         },
         {context, identifier: moduleName},
       );
@@ -834,20 +857,20 @@ export default class Runtime {
 
     const namedExports = new Set(exports);
 
-    reexports.forEach(reexport => {
+    for (const reexport of reexports) {
       if (this._resolver.isCoreModule(reexport)) {
         const exports = this.requireModule(modulePath, reexport);
         if (exports !== null && typeof exports === 'object') {
-          Object.keys(exports).forEach(namedExports.add, namedExports);
+          for (const e of Object.keys(exports)) namedExports.add(e);
         }
       } else {
         const resolved = this._resolveCjsModule(modulePath, reexport);
 
         const exports = this.getExportsOfCjs(resolved);
 
-        exports.forEach(namedExports.add, namedExports);
+        for (const e of exports) namedExports.add(e);
       }
-    });
+    }
 
     this._cjsNamedExports.set(modulePath, namedExports);
 
@@ -930,6 +953,7 @@ export default class Runtime {
       exports: {},
       filename: modulePath,
       id: modulePath,
+      isPreloading: false,
       loaded: false,
       path: path.dirname(modulePath),
     };
@@ -954,9 +978,7 @@ export default class Runtime {
 
   requireInternalModule<T = unknown>(from: string, to?: string): T {
     if (to) {
-      const require = (
-        nativeModule.createRequire ?? nativeModule.createRequireFromPath
-      )(from);
+      const require = nativeModule.createRequire(from);
       if (INTERNAL_MODULE_REQUIRE_OUTSIDE_OPTIMIZED_MODULES.has(to)) {
         return require(to);
       }
@@ -1042,6 +1064,7 @@ export default class Runtime {
         exports: {},
         filename: modulePath,
         id: modulePath,
+        isPreloading: false,
         loaded: false,
         path: path.dirname(modulePath),
       };
@@ -1212,19 +1235,19 @@ export default class Runtime {
     if (this._environment) {
       if (this._environment.global) {
         const envGlobal = this._environment.global;
-        (Object.keys(envGlobal) as Array<keyof typeof globalThis>).forEach(
-          key => {
-            const globalMock = envGlobal[key];
-            if (
-              ((typeof globalMock === 'object' && globalMock !== null) ||
-                typeof globalMock === 'function') &&
-              '_isMockFunction' in globalMock &&
-              globalMock._isMockFunction === true
-            ) {
-              globalMock.mockClear();
-            }
-          },
-        );
+        for (const key of Object.keys(envGlobal) as Array<
+          keyof typeof globalThis
+        >) {
+          const globalMock = envGlobal[key];
+          if (
+            ((typeof globalMock === 'object' && globalMock !== null) ||
+              typeof globalMock === 'function') &&
+            '_isMockFunction' in globalMock &&
+            globalMock._isMockFunction === true
+          ) {
+            globalMock.mockClear();
+          }
+        }
       }
 
       if (this._environment.fakeTimers) {
@@ -1343,6 +1366,14 @@ export default class Runtime {
     this._moduleMocker.clearAllMocks();
   }
 
+  enterTestCode(): void {
+    this.isInsideTestCode = true;
+  }
+
+  leaveTestCode(): void {
+    this.isInsideTestCode = false;
+  }
+
   teardown(): void {
     this.restoreAllMocks();
     this.resetModules();
@@ -1455,7 +1486,7 @@ export default class Runtime {
         'The first argument to require.resolve.paths must be a string. Received null or undefined.',
       );
     }
-    if (!moduleName.length) {
+    if (moduleName.length === 0) {
       throw new Error(
         'The first argument to require.resolve.paths must not be the empty string.',
       );
@@ -1485,6 +1516,11 @@ export default class Runtime {
       );
       process.exitCode = 1;
       return;
+    }
+    if (this.isInsideTestCode === false) {
+      throw new ReferenceError(
+        'You are trying to `import` a file outside of the scope of the test code.',
+      );
     }
 
     // If the environment was disposed, prevent this module from being executed.
@@ -1580,7 +1616,7 @@ export default class Runtime {
         module.path, // __dirname
         module.filename, // __filename
         lastArgs[0],
-        ...lastArgs.slice(1).filter(notEmpty),
+        ...lastArgs.slice(1).filter(isNonNullable),
       );
     } catch (error: any) {
       this.handleExecutionError(error, module);
@@ -1705,10 +1741,10 @@ export default class Runtime {
       function () {
         // @ts-expect-error: TS doesn't know what `this` is
         this.setExport('default', required);
-        Object.entries(required).forEach(([key, value]) => {
+        for (const [key, value] of Object.entries(required)) {
           // @ts-expect-error: TS doesn't know what `this` is
           this.setExport(key, value);
-        });
+        }
       },
       // should identifier be `node://${moduleName}`?
       {context, identifier: moduleName},
@@ -1791,6 +1827,7 @@ export default class Runtime {
         exports: {},
         filename,
         id: filename,
+        isPreloading: false,
         loaded: false,
         path: path.dirname(filename),
       });
@@ -1799,33 +1836,15 @@ export default class Runtime {
     // should we implement the class ourselves?
     class Module extends nativeModule.Module {}
 
-    Object.entries(nativeModule.Module).forEach(([key, value]) => {
+    for (const [key, value] of Object.entries(nativeModule.Module)) {
       // @ts-expect-error: no index signature
       Module[key] = value;
-    });
+    }
 
     Module.Module = Module;
 
     if ('createRequire' in nativeModule) {
       Module.createRequire = createRequire;
-    }
-    if ('createRequireFromPath' in nativeModule) {
-      Module.createRequireFromPath = function createRequireFromPath(
-        filename: string | URL,
-      ) {
-        if (typeof filename !== 'string') {
-          const error: NodeJS.ErrnoException = new TypeError(
-            `The argument 'filename' must be string. Received '${filename}'.${
-              filename instanceof URL
-                ? ' Use createRequire for URL filename.'
-                : ''
-            }`,
-          );
-          error.code = 'ERR_INVALID_ARG_TYPE';
-          throw error;
-        }
-        return createRequire(filename);
-      };
     }
     if ('syncBuiltinESMExports' in nativeModule) {
       // cast since TS seems very confused about whether it exists or not
@@ -2171,6 +2190,11 @@ export default class Runtime {
         );
         process.exitCode = 1;
       }
+      if (this.isInsideTestCode === false) {
+        throw new ReferenceError(
+          'You are trying to access a property or method of the Jest environment outside of the scope of the test code.',
+        );
+      }
 
       return this._fakeTimersImplementation!;
     };
@@ -2396,7 +2420,7 @@ export default class Runtime {
     const originalStack = new ReferenceError(`${errorMessage}${testPath}`)
       .stack!.split('\n')
       // Remove this file from the stack (jest-message-utils will keep one line)
-      .filter(line => line.indexOf(__filename) === -1)
+      .filter(line => !line.includes(__filename))
       .join('\n');
 
     const {message, stack} = separateMessageFromStack(originalStack);
@@ -2427,7 +2451,7 @@ export default class Runtime {
       '__filename',
       this._config.injectGlobals ? 'jest' : undefined,
       ...this._config.sandboxInjectedGlobals,
-    ].filter(notEmpty);
+    ].filter(isNonNullable);
   }
 
   private handleExecutionError(e: Error, module: Module): never {
@@ -2476,10 +2500,10 @@ export default class Runtime {
     const module = new SyntheticModule(
       Object.keys(globals),
       function () {
-        Object.entries(globals).forEach(([key, value]) => {
+        for (const [key, value] of Object.entries(globals)) {
           // @ts-expect-error: TS doesn't know what `this` is
           this.setExport(key, value);
-        });
+        }
       },
       {context, identifier: '@jest/globals'},
     );
@@ -2537,16 +2561,6 @@ export default class Runtime {
   setGlobalsForRuntime(globals: JestGlobals): void {
     this.jestGlobals = globals;
   }
-}
-
-function invariant(condition: unknown, message?: string): asserts condition {
-  if (!condition) {
-    throw new Error(message);
-  }
-}
-
-function notEmpty<T>(value: T | null | undefined): value is T {
-  return value !== null && value !== undefined;
 }
 
 async function evaluateSyntheticModule(module: SyntheticModule) {